+++ src/test/converter2/issues/gh1733.ts
@@ -1,0 +1,9 @@
+/**
+ * @typeParam T T docs
+ */
+export type Foo<T> = T;
+
+/**
+ * @typeParam T T docs
+ */
+export class Bar<T> {}